Remove detekt convention plugin. Instead apply it to all module convention plugins
