+++ android/build-logic/convention/src/main/kotlin/com/theguardian/convention/KotlinLibraryConventionPlugin.kt
@@ -3,6 +3,7 @@ package com.theguardian.convention
 import com.theguardian.convention.shared.addBaseDependencies
 import com.theguardian.convention.shared.libs
 import com.theguardian.convention.shared.plugin
+import com.theguardian.convention.shared.setupDetekt
 import org.gradle.api.Plugin
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.KotlinJvmProjectExtension
@@ -23,6 +24,8 @@ class KotlinLibraryConventionPlugin : Plugin<Project> {
                 apply("jvm-module-test")
                 apply(libs.plugin("kotlinter").pluginId)
             }
+
+            setupDetekt()
 
             addBaseDependencies<KotlinJvmProjectExtension>()
         }
