+++ android/source/build.gradle.kts
@@ -3,7 +3,6 @@ plugins {
     signing
     alias(libs.plugins.guardian.library.android)
     alias(libs.plugins.guardian.compose.library)
-    alias(libs.plugins.guardian.detekt)
 }
 
 android {
